+++ CHANGELOG.md
@@ -6,6 +6,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 1.12.65, 9/30/2020
+
+### Added
+
+N/A
+
+### Removed
+
+N/A
+
+### Changed
+
+Decouple Kafka pub/sub from kafka connector so that native pub/sub can be used when
+application is deployed without using Kafka as a service mesh. 
+
+i.e. standalone apps can talk to each other using native pub/sub.
+
+---
 ## Version 1.12.64, 9/28/2020
 
 ### Added